feat: implement useLogin hook for handling login functionality and form submission
