+++ src/sections/auth/components/login-form.tsx
@@ -1,35 +1,13 @@
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import { useAppDispatch } from "@/hooks/use-selector";
 import { cn } from "@/lib/utils";
-import { useRegisterMutation } from "@/store/services/auth/authApi";
-import { setUser } from "@/store/slices/auth/auth-slice";
-import { useState } from "react";
-import { Link, useNavigate } from "react-router";
+import { Link } from "react-router";
+import useLogin from "../hooks/use-login";
 
 export const LoginForm = () => {
-  const [email, setEmail] = useState("eddytalavera073@gmail.com");
-  const [password, setPassword] = useState("plumx34045");
-  const navigate = useNavigate();
-  const [register] = useRegisterMutation();
-  const dispatch = useAppDispatch();
-
-  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    try {
-      const data = await register({ email, password }).unwrap();
-      if (data && data.user) {
-        dispatch(setUser(data.user));
-        console.log("Registration successful:", data);
-        navigate("/");
-      } else {
-        console.error("Registration failed: No user data returned");
-      }
-    } catch (error) {
-      console.error("Registration failed:", error);
-    }
-  };
+  const { email, setEmail, password, setPassword, handleSubmit, navigate } =
+    useLogin();
 
   return (
     <form onSubmit={handleSubmit}>
